[REF] core: move orm's tests from base to test_orm
