+++ odoo/addons/base/tests/test_base.py
@@ -6,95 +6,6 @@ from odoo.tests.common import tagged, TransactionCase
 
 
 @tagged('at_install', '-post_install')  # LEGACY at_install
-class TestParentStore(TransactionCase):
-    """ Verify that parent_store computation is done right """
-
-    def setUp(self):
-        super(TestParentStore, self).setUp()
-
-        # force res_partner_category.copy() to copy children
-        category = self.env['res.partner.category']
-        self.patch(category._fields['child_ids'], 'copy', True)
-
-        # setup categories
-        self.root = category.create({'name': 'Root category'})
-        self.cat0 = category.create({'name': 'Parent category', 'parent_id': self.root.id})
-        self.cat1 = category.create({'name': 'Child 1', 'parent_id': self.cat0.id})
-        self.cat2 = category.create({'name': 'Child 2', 'parent_id': self.cat0.id})
-        self.cat21 = category.create({'name': 'Child 2-1', 'parent_id': self.cat2.id})
-
-    def test_duplicate_parent(self):
-        """ Duplicate the parent category and verify that the children have been duplicated too """
-        new_cat0 = self.cat0.copy()
-        new_struct = new_cat0.search([('parent_id', 'child_of', new_cat0.id)])
-        self.assertEqual(len(new_struct), 4, "After duplication, the new object must have the childs records")
-        old_struct = new_cat0.search([('parent_id', 'child_of', self.cat0.id)])
-        self.assertEqual(len(old_struct), 4, "After duplication, previous record must have old childs records only")
-        self.assertFalse(new_struct & old_struct, "After duplication, nodes should not be mixed")
-
-    def test_missing_parent(self):
-        """ Missing parent id should not raise an error. """
-        # Missing parent with _parent_store
-        new_cat0 = self.cat0.copy()
-        records = new_cat0.search([('parent_id', 'parent_of', 999999999)])
-        self.assertEqual(len(records), 0)
-
-        # Missing parent without _parent_store
-        category = self.env['res.partner.category']
-        self.patch(self.env.registry['res.partner.category'], '_parent_store', False)
-        records = category.search([('parent_id', 'child_of', 999999999)])
-        self.assertEqual(len(records), 0)
-
-    def test_missing_child(self):
-        """ Missing child id should not raise an error. """
-        # Missing child with _parent_store
-        new_cat0 = self.cat0.copy()
-        records = new_cat0.search([('parent_id', 'child_of', 999999999)])
-        self.assertEqual(len(records), 0)
-
-        # Missing child without _parent_store
-        category = self.env['res.partner.category']
-        self.patch(self.env.registry['res.partner.category'], '_parent_store', False)
-        records = category.search([('parent_id', 'child_of', 999999999)])
-        self.assertEqual(len(records), 0)
-
-    def test_duplicate_children_01(self):
-        """ Duplicate the children then reassign them to the new parent (1st method). """
-        new_cat1 = self.cat1.copy()
-        new_cat2 = self.cat2.copy()
-        new_cat0 = self.cat0.copy({'child_ids': []})
-        (new_cat1 + new_cat2).write({'parent_id': new_cat0.id})
-        new_struct = new_cat0.search([('parent_id', 'child_of', new_cat0.id)])
-        self.assertEqual(len(new_struct), 4, "After duplication, the new object must have the childs records")
-        old_struct = new_cat0.search([('parent_id', 'child_of', self.cat0.id)])
-        self.assertEqual(len(old_struct), 4, "After duplication, previous record must have old childs records only")
-        self.assertFalse(new_struct & old_struct, "After duplication, nodes should not be mixed")
-
-    def test_duplicate_children_02(self):
-        """ Duplicate the children then reassign them to the new parent (2nd method). """
-        new_cat1 = self.cat1.copy()
-        new_cat2 = self.cat2.copy()
-        new_cat0 = self.cat0.copy({'child_ids': [Command.set((new_cat1 + new_cat2).ids)]})
-        new_struct = new_cat0.search([('parent_id', 'child_of', new_cat0.id)])
-        self.assertEqual(len(new_struct), 4, "After duplication, the new object must have the childs records")
-        old_struct = new_cat0.search([('parent_id', 'child_of', self.cat0.id)])
-        self.assertEqual(len(old_struct), 4, "After duplication, previous record must have old childs records only")
-        self.assertFalse(new_struct & old_struct, "After duplication, nodes should not be mixed")
-
-    def test_duplicate_children_03(self):
-        """ Duplicate the children then reassign them to the new parent (3rd method). """
-        new_cat1 = self.cat1.copy()
-        new_cat2 = self.cat2.copy()
-        new_cat0 = self.cat0.copy({'child_ids': []})
-        new_cat0.write({'child_ids': [Command.link(new_cat1.id), Command.link(new_cat2.id)]})
-        new_struct = new_cat0.search([('parent_id', 'child_of', new_cat0.id)])
-        self.assertEqual(len(new_struct), 4, "After duplication, the new object must have the childs records")
-        old_struct = new_cat0.search([('parent_id', 'child_of', self.cat0.id)])
-        self.assertEqual(len(old_struct), 4, "After duplication, previous record must have old childs records only")
-        self.assertFalse(new_struct & old_struct, "After duplication, nodes should not be mixed")
-
-
-@tagged('at_install', '-post_install')  # LEGACY at_install
 class TestGroups(TransactionCase):
 
     def test_res_groups_fullname_search(self):
