+++ odoo/addons/test_orm/tests/__init__.py
@@ -1,8 +1,12 @@
 from . import (
     test_access_feedback,
     test_access_monetary_related,
+    test_acl,
     test_action_bindings,
+    test_all_models,
+    test_api,
     test_autovacuum,
+    test_cache,
     test_check_access,
     test_company_checks,
     test_config_settings,
@@ -10,6 +14,7 @@ from . import (
     test_create,
     test_convert_env,
     test_domain,
+    test_domain_expression,
     test_fields,
     test_html_converter,
     test_indexed_translation,
@@ -18,18 +23,25 @@ from . import (
     test_json_field,
     test_many2many,
     test_guess_mimetypes,
+    test_models,
     test_modules_db,
     test_one2many,
+    test_parent_store,
     test_performance,
     test_properties,
+    test_query,
     test_qweb_float,
     test_read_group_override,
     test_read_group_private,
     test_read_grouping_sets,
+    test_registry_signaling,
     test_related_translation,
     test_schema,
     test_search,
+    test_search_order,
     test_sort,
     test_timeit,
+    test_transactions,
+    # test_uninstall  # Cannot be handled by the runbot. Makes Odoo loop forever when done during an installation.
     test_views,
 )
