+++ odoo/addons/test_orm/tests/test_fields.py
@@ -17,7 +17,7 @@ from odoo.tools.image import image_data_uri
 
 from odoo.addons.base.tests.common import TransactionCaseWithUserDemo
 from odoo.addons.base.tests.files import SVG_B64, ZIP_RAW
-from odoo.addons.base.tests.test_expression import TransactionExpressionCase
+from odoo.addons.test_orm.tests.test_domain_expression import TransactionExpressionCase
 
 
 @tagged('at_install', '-post_install')  # LEGACY at_install
@@ -5336,3 +5336,34 @@ class TestAttributes(TransactionCase):
 
         with self.assertRaises(AttributeError):
             instance.unknown = 42
+
+
+class TestCompanyDependent(TransactionCase):
+    def test_orm_ondelete_restrict(self):
+        # model_A
+        #  | field_a                           company dependent many2one is
+        #  | company dependent many2one        stored as jsonb and doesn't
+        #  | (ondelete='restrict')             have db ON DELETE action
+        #  v
+        # model_B
+        #  | field_b                           if a row for model_B is deleted
+        #  | many2one (ondelete='cascade')     because of ON DELETE CASCADE,
+        #  v                                   model_A will reference a deleted
+        # model_C                              row and logically be NULL when read
+        #
+        #                                      this test asks you to move the
+        #                                      ON DELETE CASCADE logic of model_B
+        #                                      to ORM and remove ondelete='cascade'
+
+        for model in self.env.registry.values():
+            for field in model._fields.values():
+                if field.company_dependent and field.type == 'many2one' and field.ondelete.lower() == 'restrict':
+                    for comodel_field in self.env[field.comodel_name]._fields.values():
+                        self.assertFalse(
+                            comodel_field.type == 'many2one' and comodel_field.ondelete == 'cascade',
+                            (f'when a row for {comodel_field.comodel_name} is deleted, a row for {comodel_field.model_name} '
+                             f'may also be deleted for sake of on delete cascade field {comodel_field}, which will '
+                             f'bypass the ORM ondelete="restrict" check for a company dependent many2one field {field}. '
+                             f'Please override the unlink method of {comodel_field.comodel_name} and do the ORM on '
+                             f'delete cascade logic and remove/override the ondelete="cascade" of {comodel_field}')
+                        )
