+++ odoo/addons/test_orm/tests/test_search.py
@@ -1,4 +1,4 @@
-from odoo.addons.base.tests.test_expression import TransactionExpressionCase
+from odoo.addons.test_orm.tests.test_domain_expression import TransactionExpressionCase
 from odoo.fields import Command, Domain
 from odoo.tests import tagged, TransactionCase
